Move Genres component into organisms
